speed up

+++ include/animate_fluids.h
@@ -4,7 +4,7 @@
 void animate_fluids(
     Eigen::MatrixXd & X, 
     Eigen::MatrixXd & V, 
-    Eigen::MatrixXd & N,
+    Eigen::MatrixXi & N,
     const Eigen::Vector3d & low_bound,
     const Eigen::Vector3d & up_bound,
     const int numofparticles,
