+++ include/calculate_delta_p.h
@@ -3,7 +3,7 @@
 void calculate_delta_p(
     Eigen::MatrixXd & delta_p,
     const Eigen::MatrixXd & X,
-    const Eigen::MatrixXd & N,
+    const Eigen::MatrixXi & N,
     const Eigen::VectorXd & lambda,
     const double rho_0,
     const int numofparticles,
