+++ include/calculate_lambda.h
@@ -2,7 +2,7 @@
 
 void calculate_lambda(
     const Eigen::MatrixXd & x,
-    const Eigen::MatrixXd & N,
+    const Eigen::MatrixXi & N,
     Eigen::VectorXd & lambda,
     const double h,
     const double rho_0 = 1000    
@@ -11,13 +11,13 @@ void calculate_lambda(
 double W(const Eigen::Vector3d r, const double h);
 
 double C(const Eigen::MatrixXd x, 
-    const Eigen::MatrixXd N,
+    const Eigen::MatrixXi N,
     const double rho_0,
     const int i,
     const double h);
 
 double grad_C_squared(const Eigen::MatrixXd x, 
-    const Eigen::MatrixXd N,
+    const Eigen::MatrixXi N,
     const double rho_0,
     const int i,
     const double h);
