+++ src/forces.cpp
@@ -7,7 +7,7 @@
 
 void forces(
     Eigen::MatrixXd & X,
-    Eigen::MatrixXd & N,
+    Eigen::MatrixXi & N,
     const int numofparticles
     ){
